fixed brightons bullshit

+++ star-rx72n-firmware/lib/rx_frame/inc/rx_crc_internal.h
@@ -5,9 +5,35 @@
  * Provides compile-time selection between hardware and software CRC-32.
  * This header is internal to the rx_frame library.
  *
- * Configuration:
- * - Define RX_CRC32_USE_HARDWARE to use hardware CRC (default on RX target)
- * - Define RX_CRC32_USE_SOFTWARE to force software CRC (for host testing)
+ * ## Design Rationale
+ *
+ * The CRC module uses a compile-time selection strategy:
+ *
+ * 1. **Hardware CRC** (default on RX72N target):
+ *    - Uses RX72N CRC Calculator peripheral at 0x00088280
+ *    - ~10x faster than software for large buffers
+ *    - IEEE 802.3 polynomial (0x04C11DB7)
+ *    - Automatically enabled when `__RX__` is defined
+ *
+ * 2. **Software CRC** (default on host, available as fallback):
+ *    - Lookup table implementation (256-entry, 1KB table)
+ *    - Bit-exact compatible with hardware and Go's crc32.ChecksumIEEE()
+ *    - Used for host-side unit testing (no hardware available)
+ *    - Always compiled via rx_crc32_update_sw() for incremental CRC
+ *
+ * ## Why Both Implementations?
+ *
+ * - **Testing**: Host-side tests validate CRC correctness without hardware
+ * - **Debugging**: Software can be forced on target to compare results
+ * - **Incremental CRC**: rx_crc32_update_sw() is always available because
+ *   hardware state management for incremental updates is complex
+ *
+ * ## Configuration
+ *
+ * | Build Target | Default | Override |
+ * |--------------|---------|----------|
+ * | RX72N (`__RX__`) | Hardware | Define `RX_CRC32_USE_SOFTWARE` |
+ * | Host (testing) | Software | N/A |
  *
  * References:
  * - RX72N Group User's Manual: Hardware (CRC Calculator section)
@@ -21,7 +47,6 @@
 #define STAR_RX_CRC_INTERNAL_H
 
 #include <stdbool.h>
-#include <stddef.h>
 #include <stdint.h>
 
 #include "rx_err.h"
@@ -33,11 +58,22 @@ extern "C" {
 /* =============================================================================
  * Configuration - Compile-time Hardware/Software Selection
  *
- * Default behavior:
- * - On RX target (__RX__ defined): Use hardware CRC
- * - On host (no __RX__): Use software CRC for testing
+ * This preprocessor logic implements the following decision matrix:
  *
- * Override by defining RX_CRC32_USE_SOFTWARE before including this header.
+ * | __RX__ | RX_CRC32_USE_SOFTWARE | Result              |
+ * |--------|----------------------|---------------------|
+ * | Yes    | No                   | Hardware (default)  |
+ * | Yes    | Yes                  | Software (override) |
+ * | No     | (any)                | Software (testing)  |
+ *
+ * To force software CRC on hardware target:
+ *   gcc -DRX_CRC32_USE_SOFTWARE ...
+ *   or #define RX_CRC32_USE_SOFTWARE before including this header
+ *
+ * The software implementation is always compiled (rx_crc32_sw.c) because:
+ * 1. It's needed for host-side unit tests
+ * 2. rx_crc32_update_sw() is used for incremental CRC on hardware targets
+ * 3. It enables A/B comparison testing on hardware for validation
  * =============================================================================
  */
 
@@ -85,7 +121,7 @@ rx_err_t rx_crc_deinit(void);
  * @param[in] len  Data length in bytes
  * @return CRC-32 checksum (0 if data is NULL or len is 0)
  */
-uint32_t rx_crc32_ieee_impl(const uint8_t *data, size_t len);
+uint32_t rx_crc32_ieee_impl(const uint8_t* data, uint32_t len);
 
 /**
  * @brief Update CRC-32 with additional data (implementation)
@@ -97,7 +133,7 @@ uint32_t rx_crc32_ieee_impl(const uint8_t *data, size_t len);
  * @param[in] len  Data length in bytes
  * @return Updated CRC-32 checksum
  */
-uint32_t rx_crc32_update_impl(uint32_t crc, const uint8_t *data, size_t len);
+uint32_t rx_crc32_update_impl(uint32_t crc, const uint8_t* data, uint32_t len);
 
 /**
  * @brief Software CRC-32 update (always available)
@@ -110,7 +146,7 @@ uint32_t rx_crc32_update_impl(uint32_t crc, const uint8_t *data, size_t len);
  * @param[in] len  Data length in bytes
  * @return Updated CRC-32 checksum
  */
-uint32_t rx_crc32_update_sw(uint32_t crc, const uint8_t *data, size_t len);
+uint32_t rx_crc32_update_sw(uint32_t crc, const uint8_t* data, uint32_t len);
 
 #ifdef __cplusplus
 }
